Supported all entry point return types.

+++ ArkeCLR.Runtime/Execution/IExecutionEngine.cs
@@ -4,6 +4,6 @@ using System.Collections.Generic;
 
 namespace ArkeCLR.Runtime.Execution {
     public interface IExecutionEngine {
-        int Run(Assembly entryAssembly, IReadOnlyCollection<Assembly> references, Action<string> logger);
+        long Run(Assembly entryAssembly, IReadOnlyCollection<Assembly> references, Action<string> logger);
     }
 }
